Create backup_retention.ps1
Let me explain what the script does:

It sets the path to the folder you want to clean up and the number of days to keep files.
It gets the current date and uses it to calculate the date threshold for files older than $daysToKeep days.
It gets a list of files older than the threshold using Get-ChildItem and Where-Object.
It compresses each old file using Compress-Archive and adds a .zip extension to the file name.
It deletes the original files using Remove-Item.
Finally, it deletes the compressed files older than $daysToKeep days using Get-ChildItem and Remove-Item again.
Note: Make sure to update the $folderPath variable to the actual path you want to clean up. Also, be careful when running this script, as it permanently deletes files!