backup-folder.sh
Here is a bash script that performs a backup with the following features:

Specifies a source and destination directory
Compresses the backup using tar and gzip
Names the backup file with the current date
Verifies the integrity of the files using md5sum before compressing
Sends an email to a specified recipient with the status of the execution

This script uses find and md5sum to create hash files for the source and destination directories, and then compares them to ensure the integrity of the files. It then creates a backup file using tar and gzip, and verifies its integrity using md5sum. Finally, it sends an email to the specified recipient with the status of the execution.

Please note that you should replace the placeholders (/path/to/source/directory, /path/to/destination/directory, example@example.com) with the actual values for your environment. Also, make sure to test the script in a non-production environment before running it in production.